extract BackInTimeDebugService as an interface
