+++ backintime-runtime/src/androidMain/kotlin/com/github/kitakkun/backintime/runtime/flipper/BackInTimeFlipperPlugin.kt
@@ -2,12 +2,12 @@ package com.github.kitakkun.backintime.runtime.flipper
 
 import com.facebook.flipper.core.FlipperConnection
 import com.facebook.flipper.core.FlipperPlugin
-import com.github.kitakkun.backintime.runtime.BackInTimeDebugService
+import com.github.kitakkun.backintime.runtime.BackInTimeDebugServiceImpl
 import com.github.kitakkun.backintime.runtime.connector.BackInTimeConnector
 
 class BackInTimeFlipperPlugin : FlipperPlugin {
     private var connector: BackInTimeConnector? = null
-    private val service = BackInTimeDebugService
+    private val service = BackInTimeDebugServiceImpl
 
     override fun getId(): String = "back-in-time"
     override fun runInBackground(): Boolean = true
