+++ test/src/commonTest/kotlin/com/github/kitakkun/backintime/test/base/BackInTimeDebugServiceTest.kt
@@ -1,13 +1,13 @@
 package com.github.kitakkun.backintime.test.base
 
-import com.github.kitakkun.backintime.runtime.BackInTimeDebugService
+import com.github.kitakkun.backintime.runtime.BackInTimeDebugServiceImpl
 import com.github.kitakkun.backintime.websocket.event.BackInTimeDebugServiceEvent
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 
 abstract class BackInTimeDebugServiceTest {
     private val mockConnector: MockConnector = MockConnector()
-    private val service = BackInTimeDebugService
+    private val service = BackInTimeDebugServiceImpl
 
     private val events get() = mockConnector.eventsFromService
     protected val registerInstanceEvents get() = events.filterIsInstance<BackInTimeDebugServiceEvent.RegisterInstance>()
